pyproject.toml
[project]
name = "ticket-service-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "fastapi>=0.121.0",
    "fastapi-mcp>=0.4.0",
    "hypercorn>=0.17.3",
    "httpx>=0.27.0",
    "mcp>=1.20.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.38.0",
    "letta-client>=0.1.324",
    "sqlalchemy>=2.0.31",
    "alembic>=1.13.2",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
]